+++ arangod/Aql/ExecutionNode/JoinNode.cpp
@@ -41,6 +41,7 @@
 #include "Aql/RegisterPlan.h"
 #include "Aql/SingleRowFetcher.h"
 #include "Basics/AttributeNameParser.h"
+#include "Basics/StaticStrings.h"
 #include "Basics/StringUtils.h"
 #include "Basics/VelocyPackHelper.h"
 #include "Containers/FlatHashMap.h"
