+++ arangod/Aql/OptimizerUtils.cpp
@@ -44,6 +44,7 @@
 #include "Aql/RegisterPlan.h"
 #include "Aql/SortCondition.h"
 #include "Aql/Variable.h"
+#include "Basics/StaticStrings.h"
 #include "IResearch/IResearchFeature.h"
 #include "Indexes/Index.h"
 #include "Logger/LogMacros.h"
