+++ arangod/IResearch/IResearchFilterContext.h
@@ -25,7 +25,6 @@
 
 #include <string_view>
 
-#include "IResearch/IResearchInvertedIndexMeta.h"
 #include "IResearch/IResearchLinkMeta.h"
 #include "IResearch/IResearchFilterOptimization.h"
 #include "IResearch/IResearchInvertedIndexMeta.h"
