+++ arangod/Network/Methods.h
@@ -30,6 +30,7 @@
 #include "GeneralServer/RequestLane.h"
 #include "Inspection/VPack.h"
 #include "Network/ConnectionPool.h"
+#include "Network/RequestOptions.h"
 #include "Network/types.h"
 
 #include <fuerte/message.h>
@@ -132,45 +133,6 @@ struct Response {
 
 static_assert(std::is_nothrow_move_constructible<Response>::value, "");
 using FutureRes = arangodb::futures::Future<Response>;
-
-static constexpr Timeout TimeoutDefault = Timeout(120.0);
-
-// Container for optional (often defaulted) parameters
-struct RequestOptions {
-  std::string database;
-  std::string contentType;  // uses vpack by default
-  std::string acceptType;   // uses vpack by default
-  fuerte::StringMap parameters;
-  Timeout timeout = TimeoutDefault;
-  // retry if answer is "datasource not found"
-  bool retryNotFound = false;
-  // do not use Scheduler queue
-  bool skipScheduler = false;
-  // send x-arango-hlc header with outgoing request, so that that peer can
-  // update its own HLC value to at least the value of our HLC
-  bool sendHLCHeader = true;
-  // transparently handle content-encoding. enabling this will automatically
-  // uncompress responses that have the `Content-Encoding: gzip|deflate` header
-  // set.
-  bool handleContentEncoding = true;
-  // allow to compress the request
-  bool allowCompression = true;
-  RequestLane continuationLane = RequestLane::CONTINUATION;
-
-  // Normally this is empty, if it is set to the ID of a server in the
-  // cluster, we will direct a read operation to a shard not as usual to
-  // the leader, but rather to the server given here. This is read for
-  // the "allowDirtyReads" options when we want to read from followers.
-  std::string overrideDestination;
-
-  template<typename K, typename V>
-  RequestOptions& param(K&& key, V&& val) {
-    TRI_ASSERT(!std::string_view{val}.empty());  // cannot parse it on receiver
-    this->parameters.insert_or_assign(std::forward<K>(key),
-                                      std::forward<V>(val));
-    return *this;
-  }
-};
 
 /// @brief send a request to a given destination
 /// This method must not throw under penalty of ...
